BasicBank: use similar algo for mapping storage
The storage slot should be:
  0x00: value of variable slot
  0x20: address

and the two words are hashed to form the storage slot for
balance[address] where balance.slot is a number.

This allows for multiple mappings in a module, should it be
necessary; which it isn't here.

This update increases the executable byte size from 238 to 256